+++ include/ev/fs.h
@@ -205,10 +205,10 @@ struct ev_fs_req_s
 
     union
     {
-        ev_fs_stat_t            fileinfo;       /**< File information */
+        ev_fs_stat_t*           stat;           /**< File information */
         ev_list_t               dirents;        /**< Dirent list */
         ev_buf_t                filecontent;    /**< File content */
-    }rsp;
+    } rsp;
 };
 
 #define EV_FS_REQ_INVALID \
@@ -220,7 +220,7 @@ struct ev_fs_req_s
         NULL,\
         EV_EINPROGRESS,\
         { { NULL, 0, 0 } },\
-        { EV_FS_STAT_INVALID },\
+        { NULL },\
     }
 
 /**
@@ -354,16 +354,7 @@ EV_API ssize_t ev_file_pwrite(ev_file_t* file, ev_fs_req_t* req, ev_buf_t bufs[]
  * @param[in] cb        Result callback.
  * @return              #ev_errno_t
  */
-EV_API int ev_file_stat(ev_file_t* file, ev_fs_req_t* req, ev_file_cb cb);
-
-/**
- * @brief Like #ev_file_stat(), but work in synchronous mode.
- * @see ev_file_stat()
- * @param[in] file      File handle.
- * @param[out] stat     File status.
- * @return              #ev_errno_t
- */
-EV_API int ev_file_stat_sync(ev_file_t* file, ev_fs_stat_t* stat);
+EV_API int ev_file_stat(ev_file_t* file, ev_fs_req_t* req, ev_fs_stat_t* stat, ev_file_cb cb);
 
 /**
  * @brief Get all entry in directory.