+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 4. merge `ev_file_pread_sync()` with `ev_file_pread()`
 5. merge `ev_file_write_sync()` with `ev_file_write()`
 6. merge `ev_file_pwrite_sync()` with `ev_file_pwrite()`
+7. merge `ev_file_stat_sync()` with `ev_file_stat()`
 
 ### Bug Fixes
 1. `ev_hrtime()` no longer require initialize event loop first.